Replace git usage w/run_cmd only, relocate/share git ops
